+++ app/src/main/AndroidManifest.xml
@@ -16,19 +16,19 @@
             android:name=".checklist.InformationActivity"
             android:exported="false" />
         <activity
-            android:name=".MoneyActivity"
+            android:name=".settings.MoneyActivity"
             android:exported="false" />
         <activity
             android:name=".changeposition.ChangePositionActivity"
             android:exported="false" />
         <activity
-            android:name=".checklist.FamilyEditActivity"
+            android:name=".checklist.edit.FamilyEditActivity"
             android:exported="false" />
         <activity
             android:name=".checklist.edit.EditActivity"
             android:exported="false" />
         <activity
-            android:name=".SettingActivity"
+            android:name=".settings.SettingActivity"
             android:exported="false" />
         <activity
             android:name=".add.AddActivity"
