+++ app/src/main/java/com/lostark/lostarkassistanthomework/MainViewModel.kt
@@ -1,0 +1,6 @@
+package com.lostark.lostarkassistanthomework
+
+import androidx.lifecycle.ViewModel
+
+class MainViewModel : ViewModel() {
+}
