+++ app/src/main/java/com/lostark/lostarkassistanthomework/checklist/ChecklistViewModel.kt
@@ -1,0 +1,6 @@
+package com.lostark.lostarkassistanthomework.checklist
+
+import androidx.lifecycle.ViewModel
+
+class ChecklistViewModel : ViewModel() {
+}
