+++ app/src/main/java/com/lostark/lostarkassistanthomework/checklist/edit/FamilyEditViewModel.kt
@@ -1,0 +1,6 @@
+package com.lostark.lostarkassistanthomework.checklist.edit
+
+import androidx.lifecycle.ViewModel
+
+class FamilyEditViewModel : ViewModel() {
+}
