+++ app/src/main/java/com/lostark/lostarkassistanthomework/settings/ChracterRecylerAdapter.kt
@@ -1,4 +1,4 @@
-package com.lostark.lostarkassistanthomework
+package com.lostark.lostarkassistanthomework.settings
 
 import android.content.Context
 import android.view.LayoutInflater
@@ -9,49 +9,40 @@ import android.widget.ImageView
 import android.widget.LinearLayout
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
-import com.lostark.lostarkassistanthomework.checklist.rooms.Homework
+import com.lostark.lostarkassistanthomework.R
+import com.lostark.lostarkassistanthomework.objects.Chracter
 
-class GoldRecylerAdapter(
-    private val list: ArrayList<Homework>,
+class ChracterRecylerAdapter(
+    private val list: ArrayList<Chracter>,
     private val context: Context,
     private val txtContent: TextView
-) : RecyclerView.Adapter<GoldRecylerAdapter.ViewHolder>() {
-    var goldcount = 0
-
-    fun syncGold() {
-        list.forEach { item ->
-            if (item.isGold) {
-                goldcount++
-            }
-        }
-    }
+) : RecyclerView.Adapter<ChracterRecylerAdapter.ViewHolder>() {
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
         val view = LayoutInflater.from(parent.context).inflate(R.layout.item_chracter_list, parent, false)
-        return GoldRecylerAdapter.ViewHolder(view, list)
+        return ViewHolder(view, list, txtContent)
     }
 
     override fun onBindViewHolder(holder: ViewHolder, position: Int) {
         val item = list[position]
         val listener = View.OnClickListener { it ->
-            if (item.isGold) {
-                item.isGold = false
+            if (item.isChecked) {
+                item.isChecked = false
                 holder.chkCheck.isChecked = false
-                goldcount--
                 holder.layoutBackground.setBackgroundResource(R.drawable.unselected_style)
             } else {
-                if (goldcount >= 6) {
-                    val toast = CustomToast(context)
-                    toast.createToast("이미 지정할 수 있는 횟수를 초과하였습니다.", false)
-                    toast.show()
-                    return@OnClickListener
-                }
-                item.isGold = true
+                item.isChecked = true
                 holder.chkCheck.isChecked = true
-                goldcount++
                 holder.layoutBackground.setBackgroundResource(R.drawable.selected_style)
             }
-            txtContent.text = "최대 6개 캐릭터 중 ${goldcount}개 캐릭터 지정됨"
+            var count = 0
+            list.forEach { it ->
+                if (it.isChecked) {
+                    count++
+                }
+            }
+            txtContent.text = "총 ${list.size}개 중 ${count}개의 캐릭터를 선택하였습니다"
+            notifyDataSetChanged()
         }
         holder.apply {
             bind(listener, item, context)
@@ -61,7 +52,7 @@ class GoldRecylerAdapter(
 
     override fun getItemCount(): Int = list.size
 
-    class ViewHolder(v: View, private val list: ArrayList<Homework>): RecyclerView.ViewHolder(v) {
+    class ViewHolder(v : View, private val list: ArrayList<Chracter>, private val txtContent: TextView) : RecyclerView.ViewHolder(v) {
         private var view: View = v
         lateinit var imgJob: ImageView
         lateinit var txtName: TextView
@@ -70,8 +61,7 @@ class GoldRecylerAdapter(
         lateinit var txtJob: TextView
         lateinit var chkCheck: CheckBox
         lateinit var layoutBackground: LinearLayout
-
-        fun bind(listener: View.OnClickListener, item: Homework, context: Context) {
+        fun bind(listener: View.OnClickListener, item: Chracter, context: Context) {
             imgJob = view.findViewById(R.id.imgJob)
             txtName = view.findViewById(R.id.txtName)
             txtServer = view.findViewById(R.id.txtServer)
@@ -87,14 +77,32 @@ class GoldRecylerAdapter(
             txtServer.text = item.server
             txtLevel.text = "Lv.${item.level.toString()}"
             txtJob.text = item.job
-            chkCheck.isChecked = item.isGold
+            chkCheck.isChecked = item.isChecked
             if (chkCheck.isChecked) {
                 layoutBackground.setBackgroundResource(R.drawable.selected_style)
             } else {
                 layoutBackground.setBackgroundResource(R.drawable.unselected_style)
             }
 
-            chkCheck.setOnClickListener(listener)
+            chkCheck.setOnClickListener {
+                if (item.isChecked) {
+                    item.isChecked = false
+                    chkCheck.isChecked = false
+                    layoutBackground.setBackgroundResource(R.drawable.unselected_style)
+                } else {
+                    item.isChecked = true
+                    chkCheck.isChecked = true
+                    layoutBackground.setBackgroundResource(R.drawable.selected_style)
+                }
+                var count = 0
+                list.forEach { it ->
+                    if (it.isChecked) {
+                        count++
+                    }
+                }
+                txtContent.text = "총 ${list.size}개 중 ${count}개의 캐릭터를 선택하였습니다"
+            }
+
             view.setOnClickListener(listener)
         }
     }
