+++ app/src/main/java/com/lostark/lostarkassistanthomework/settings/SettingViewModel.kt
@@ -1,0 +1,6 @@
+package com.lostark.lostarkassistanthomework.settings
+
+import androidx.lifecycle.ViewModel
+
+class SettingViewModel : ViewModel() {
+}
