+++ app/src/main/res/drawable/background_spinner.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <shape xmlns:android="http://schemas.android.com/apk/res/android">
     <solid android:color="#00000000" />
-    <stroke android:color="@color/stroke" android:width="1dp" />
-    <corners android:radius="15dp"/>
+    <stroke android:color="@color/strong_stroke" android:width="1dp" />
 </shape>
