+++ app/src/main/res/drawable/button_important_style.xml
@@ -2,20 +2,17 @@
 <selector xmlns:android="http://schemas.android.com/apk/res/android">
     <item android:state_enabled="false">
         <shape>
-            <corners android:radius="15dp"/>
             <solid android:color="@color/disabled"/>
             <stroke android:width="1dp" android:color="@color/disabled" />
         </shape>
     </item>
     <item android:state_pressed="true">
         <shape>
-            <corners android:radius="15dp"/>
             <solid android:color="@color/important_clicked"/>
         </shape>
     </item>
     <item android:state_pressed="false">
         <shape>
-            <corners android:radius="15dp"/>
             <solid android:color="@color/important"/>
         </shape>
     </item>
