+++ app/src/main/res/layout/activity_main.xml
@@ -1,34 +1,43 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
+<layout
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@color/background"
-    android:orientation="vertical"
-    tools:context=".MainActivity">
-    <androidx.appcompat.widget.Toolbar
+    xmlns:tools="http://schemas.android.com/tools">
+    <LinearLayout
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:id="@+id/toolBar"
-        android:minHeight="?attr/actionBarSize"
-        app:menu="@menu/main_menu"
-        android:background="?attr/colorPrimary"/>
-    <FrameLayout
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:layout_weight="1"
-        android:id="@+id/layoutFrame"/>
-    <com.google.android.material.bottomnavigation.BottomNavigationView
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:background="#00000000"
-        android:outlineProvider="none"
-        app:menu="@menu/main_tab"
-        app:itemIconTint="@drawable/bottom_navigation_select_text_style"
-        app:itemTextColor="@drawable/bottom_navigation_select_text_style"
-        android:layout_marginBottom="5dp"
-        android:visibility="gone"
-        android:id="@+id/bottomNavigationView"/>
-</LinearLayout>
+        android:layout_height="match_parent"
+        android:background="@color/background"
+        android:orientation="vertical"
+        tools:context=".MainActivity">
+        <androidx.appcompat.widget.Toolbar
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:id="@+id/toolBar"
+            android:minHeight="?attr/actionBarSize"
+            app:menu="@menu/main_menu"
+            android:background="?attr/colorPrimary"/>
+        <androidx.constraintlayout.widget.ConstraintLayout
+            android:layout_width="match_parent"
+            android:layout_height="match_parent">
+            <FrameLayout
+                android:layout_width="match_parent"
+                android:layout_height="0dp"
+                app:layout_constraintVertical_weight="1"
+                app:layout_constraintTop_toTopOf="parent"
+                app:layout_constraintBottom_toTopOf="@id/bottomNavigationView"
+                android:id="@+id/layoutFrame"/>
+            <com.google.android.material.bottomnavigation.BottomNavigationView
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                app:layout_constraintBottom_toBottomOf="parent"
+                android:background="#00000000"
+                android:outlineProvider="none"
+                app:menu="@menu/main_tab"
+                app:itemIconTint="@drawable/bottom_navigation_select_text_style"
+                app:itemTextColor="@drawable/bottom_navigation_select_text_style"
+                android:layout_marginBottom="5dp"
+                android:visibility="gone"
+                android:id="@+id/bottomNavigationView"/>
+        </androidx.constraintlayout.widget.ConstraintLayout>
+    </LinearLayout>
+</layout>
