+++ app/src/main/res/layout/activity_money.xml
@@ -1,53 +1,66 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    android:background="@color/background"
-    tools:context=".MoneyActivity">
-    <androidx.appcompat.widget.Toolbar
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:id="@+id/toolBar"
-        android:minHeight="?attr/actionBarSize"
-        android:background="?attr/colorPrimary"/>
+    xmlns:tools="http://schemas.android.com/tools">
+    <data>
+        <variable
+            name="moneyViewModel"
+            type="com.lostark.lostarkassistanthomework.settings.MoneyViewModel" />
+    </data>
     <LinearLayout
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="horizontal"
-        android:padding="10dp">
-        <ImageView
-            android:layout_width="20dp"
-            android:layout_height="20dp"
-            android:src="@drawable/icon_gold"
-            android:layout_gravity="center"/>
-        <TextView
-            android:layout_width="wrap_content"
+        android:layout_height="match_parent"
+        android:orientation="vertical"
+        android:background="@color/background"
+        tools:context=".settings.MoneyActivity">
+        <androidx.appcompat.widget.Toolbar
+            android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:id="@+id/txtContent"
-            android:text="최대 6개 캐릭터 중 0개 캐릭터 지정됨"
-            android:textSize="16dp"
-            android:textColor="@color/main_font"
-            android:layout_marginLeft="5dp"/>
+            android:id="@+id/toolBar"
+            android:minHeight="?attr/actionBarSize"
+            android:background="?attr/colorPrimary"/>
+        <androidx.constraintlayout.widget.ConstraintLayout
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:padding="10dp">
+            <ImageView
+                android:layout_width="20dp"
+                android:layout_height="20dp"
+                app:layout_constraintTop_toTopOf="parent"
+                app:layout_constraintLeft_toLeftOf="parent"
+                app:layout_constraintBottom_toTopOf="@id/listView"
+                android:id="@+id/imgGoldIcon"
+                android:src="@drawable/icon_gold"/>
+            <TextView
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:id="@+id/txtContent"
+                android:text="최대 6개 캐릭터 중 0개 캐릭터 지정됨"
+                android:textSize="16dp"
+                android:textColor="@color/main_font"
+                app:layout_constraintLeft_toRightOf="@id/imgGoldIcon"
+                app:layout_constraintTop_toTopOf="parent"
+                android:layout_marginLeft="5dp"/>
+            <androidx.recyclerview.widget.RecyclerView
+                android:layout_width="match_parent"
+                android:layout_height="0dp"
+                app:layout_constraintVertical_weight="1"
+                app:layout_constraintTop_toBottomOf="@id/txtContent"
+                app:layout_constraintBottom_toTopOf="@id/btnApply"
+                android:id="@+id/listView"
+                android:layout_margin="10dp"
+                app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"/>
+            <android.widget.Button
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                app:layout_constraintBottom_toBottomOf="parent"
+                android:id="@+id/btnApply"
+                android:background="@drawable/button_style"
+                android:text="적용"
+                android:textSize="14dp"
+                android:textColor="@color/main_font"
+                android:layout_margin="10dp"
+                style="?android:attr/borderlessButtonStyle"/>
+        </androidx.constraintlayout.widget.ConstraintLayout>
     </LinearLayout>
-    <androidx.recyclerview.widget.RecyclerView
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:layout_weight="1"
-        android:id="@+id/listView"
-        android:layout_marginLeft="10dp"
-        android:layout_marginRight="10dp"
-        app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"/>
-    <android.widget.Button
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:id="@+id/btnApply"
-        android:background="@drawable/button_style"
-        android:text="적용"
-        android:textSize="14dp"
-        android:textColor="@color/main_font"
-        android:layout_margin="10dp"
-        style="?android:attr/borderlessButtonStyle"/>
-</LinearLayout>
+</layout>
