+++ app/src/main/res/layout/fragment_auto.xml
@@ -1,49 +1,55 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:orientation="vertical"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
-    <LinearLayout
-        android:layout_width="match_parent"
+    <android.widget.Button
+        android:layout_width="0dp"
         android:layout_height="wrap_content"
-        android:layout_marginBottom="10dp"
-        android:orientation="horizontal">
-        <android.widget.Button
-            android:layout_width="0dp"
-            android:layout_weight="3"
-            android:layout_height="wrap_content"
-            android:id="@+id/btnSearch"
-            android:background="@drawable/button_style"
-            android:text="캐릭터 정보 불러오기"
-            android:textSize="14dp"
-            android:textColor="@color/main_font"
-            android:layout_marginRight="5dp"
-            style="?android:attr/borderlessButtonStyle"/>
-        <android.widget.Button
-            android:layout_width="0dp"
-            android:layout_weight="1"
-            android:layout_height="wrap_content"
-            android:id="@+id/btnAll"
-            android:background="@drawable/button_style"
-            android:text="모두 선택"
-            android:textSize="14dp"
-            android:textColor="@color/main_font"
-            android:layout_marginLeft="5dp"
-            style="?android:attr/borderlessButtonStyle"/>
-    </LinearLayout>
+        app:layout_constraintHorizontal_weight="3"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintLeft_toLeftOf="parent"
+        app:layout_constraintRight_toLeftOf="@id/btnAll"
+        android:drawableRight="10dp"
+        android:id="@+id/btnSearch"
+        android:background="@drawable/button_style"
+        android:text="캐릭터 정보 불러오기"
+        android:textSize="14dp"
+        android:textColor="@color/main_font"
+        android:layout_marginRight="5dp"
+        style="?android:attr/borderlessButtonStyle"/>
+    <android.widget.Button
+        android:layout_width="0dp"
+        android:layout_height="wrap_content"
+        app:layout_constraintHorizontal_weight="1"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintRight_toRightOf="parent"
+        app:layout_constraintLeft_toRightOf="@id/btnSearch"
+        android:id="@+id/btnAll"
+        android:background="@drawable/button_style"
+        android:text="모두 선택"
+        android:textSize="14dp"
+        android:textColor="@color/main_font"
+        android:layout_marginLeft="5dp"
+        style="?android:attr/borderlessButtonStyle"/>
     <TextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
+        app:layout_constraintTop_toBottomOf="@id/btnSearch"
+        app:layout_constraintLeft_toLeftOf="parent"
         android:id="@+id/txtContent"
         android:text="content"
         android:textSize="14dp"
-        android:layout_marginBottom="10dp"
+        android:layout_marginTop="10dp"
         android:textColor="@color/main_font"
         android:visibility="gone"/>
     <androidx.recyclerview.widget.RecyclerView
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
+        android:layout_height="0dp"
+        app:layout_constraintVertical_weight="1"
+        app:layout_constraintTop_toBottomOf="@id/txtContent"
+        app:layout_constraintBottom_toBottomOf="parent"
         android:id="@+id/listAuto"
+        android:layout_marginTop="10dp"
         app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"/>
-</LinearLayout>
+</androidx.constraintlayout.widget.ConstraintLayout>
