+++ app/src/main/res/values-night/colors.xml
@@ -16,7 +16,7 @@
     <color name="complete_background">#555555</color>
     <color name="stroke">#80777777</color>
     <color name="actionbar">#333333</color>
-    <color name="progress_background">#666666</color>
+    <color name="progress_background">#777777</color>
     <color name="progress">#788FE4</color>
     <color name="line_style">#444444</color>
     <color name="disabled">#FF8888</color>
